move signup and login into one component
